refactor: extract snap, rlp & e2store functionality into mods/files
